// src/FE/src/api/generated/models/documentAttachmentInfoResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Workslip.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface DocumentAttachmentInfoResponse {
  id: string;
  documentId: string;
  fileName: string;
  contentType: string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  sizeBytes: number | string;
  createdAt: string;
  /** @nullable */
  uploadedByUserId: string | null;
  /** @nullable */
  uploadedByDisplayName: string | null;
}